backend/requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
arxiv==2.2.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
cbor==1.0.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.3
dataclasses-json==0.6.7
datasets==4.0.0
debugpy==1.8.15
decorator==5.2.1
dill==0.3.8
distro==1.9.0
elastic-transport==8.17.1
elasticsearch==8.18.1
et-xmlfile==2.0.0
executing==2.2.0
fastapi==0.116.1
fastapi-mcp==0.4.0
feedparser==6.0.11
filelock==3.18.0
flagembedding==1.3.5
frozenlist==1.7.0
fsspec==2025.3.0
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
groq==0.30.0
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.1
idna==3.10
ijson==3.4.0
inscriptis==2.6.0
ipykernel==6.30.0
ipython==9.4.0
ipython-pygments-lexers==1.1.1
ir-datasets==0.5.11
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.8.1
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-elasticsearch==0.3.2
langchain-google-community==2.0.7
langchain-groq==0.3.6
langchain-mcp-adapters==0.1.9
langchain-ollama==0.3.6
langchain-tavily==0.2.11
langchain-text-splitters==0.3.9
langgraph==0.5.4
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.74
langsmith==0.4.8
lxml==6.0.0
lz4==4.4.4
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.12.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.3.2
oauthlib==3.3.1
ollama==0.5.1
openpyxl==3.1.5
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
peft==0.16.0
pillow==11.3.0
platformdirs==4.3.8
prompt-toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pure-eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentence-transformers==5.0.0
sentencepiece==0.2.0
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
simsimd==6.5.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
sse-starlette==2.4.1
stack-data==0.6.3
starlette==0.47.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.2
tomli==2.2.1
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.0
trec-car-tools==2.6
typer==0.16.0
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
unlzw3==0.2.3
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wcwidth==0.2.13
wikipedia==1.4.0
xxhash==3.5.0
yarl==1.20.1
zlib-state==0.1.9
zstandard==0.23.0
